Agent-Python/pyproject.toml
[project]
name = "lazyagent"
version = "0.1.0"
description = "AI agent with LangChain and OpenAI"
readme = "README.md"
requires-python = ">=3.9,<4.0"

dependencies = [
    "duckduckgo-search==8.1.1",
    "langchain==0.3.26",
    "langchain-community==0.3.27",
    "langchain-experimental==0.3.4",
    "langchain-openai==0.3.27",
    "openai>=1.86.0,<2.0.0",
    "python-dotenv>=1.1.1",
    "uvicorn[standard]>=0.29.0",
    "fastapi>=0.111.0",
    "pydantic>=2.0.0",
    "httpx>=0.27.0",
]
